fix: (2.36) Handle lazy exception for program rule notifications (#7601) (#7603)
* fix: Handle lazy exception for program rule notifications (#7601)

JIRA # [DHIS2-10396]

* Use Transactional at method level
