+++ dhis-2/dhis-e2e-test/src/main/resources/config.properties
@@ -13,4 +13,4 @@ user.super.password=Test1212?
 user.admin.username=taadmin
 user.admin.password=Test1212?
 
-test.cleanup=false
+test.cleanup=true
