+++ dhis-2/dhis-e2e-test/src/test/java/org/hisp/dhis/tracker/importer/RuleEngineTests.java
@@ -33,6 +33,7 @@ import org.hamcrest.Matchers;
 import org.hisp.dhis.ApiTest;
 import org.hisp.dhis.Constants;
 import org.hisp.dhis.actions.LoginActions;
+import org.hisp.dhis.actions.MessageConversationsActions;
 import org.hisp.dhis.actions.RestApiActions;
 import org.hisp.dhis.actions.metadata.MetadataActions;
 import org.hisp.dhis.actions.metadata.ProgramStageActions;
@@ -42,6 +43,7 @@ import org.hisp.dhis.dto.TrackerApiResponse;
 import org.hisp.dhis.helpers.JsonObjectBuilder;
 import org.hisp.dhis.helpers.QueryParamsBuilder;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -65,15 +67,25 @@ public class RuleEngineTests
 
     private TrackerActions trackerActions;
 
+    private LoginActions loginActions;
+
+    private MessageConversationsActions messageConversationsActions;
+
     @BeforeAll
     public void beforeAll()
     {
-
+        messageConversationsActions = new MessageConversationsActions();
         trackerActions = new TrackerActions();
+        loginActions = new LoginActions();
 
-        new LoginActions().loginAsSuperUser();
+        loginActions.loginAsSuperUser();
         new MetadataActions()
             .importAndValidateMetadata( new File( "src/test/resources/tracker/programs_with_program_rules.json" ) );
+    }
+
+    @BeforeEach
+    public void beforeEach() {
+        loginActions.loginAsSuperUser();
     }
 
     @ParameterizedTest
@@ -177,9 +189,9 @@ public class RuleEngineTests
             .body( "dataValues.value", contains( "AUTO_ASSIGNED_COMMENT" ) );
     }
 
-    @Disabled( "bug DHIS2-10127" )
     @Test
-    public void shouldSendNotification()
+    public void shouldSendProgramRuleNotification()
+        throws InterruptedException
     {
         JsonObject payload = trackerActions.buildEvent( Constants.ORG_UNIT_IDS[0], eventProgramId, "Mt6Ac5brjoK" );
 
@@ -195,17 +207,22 @@ public class RuleEngineTests
                 .addProperty( "value", "40" )
                 .build() );
 
+        loginActions.loginAsAdmin();
         ApiResponse response = new RestApiActions( "/messageConversations" ).get( "", new QueryParamsBuilder().add( "fields=*" ) );
 
         int size = response.getBody().getAsJsonArray( "messageConversations" ).size();
 
+        loginActions.loginAsSuperUser();
+
         trackerActions.postAndGetJobReport( payload )
             .validateSuccessfulImport();
 
-        new RestApiActions( "/messageConversations?fields=*" ).get( "", new QueryParamsBuilder().add( "fields=*" ) )
+        loginActions.loginAsAdmin();
+        messageConversationsActions.waitForNotification( size + 1 );
+        messageConversationsActions.get( "", new QueryParamsBuilder().add( "fields=*" ) )
             .validate()
             .statusCode( 200 )
-            .body( "messageConversations", arrayWithSize( size + 1 ) )
+            .body( "messageConversations", hasSize( size + 1 ) )
             .body( "messageConversations.subject", hasItem( "Program rule triggered" ) );
     }
